# api/openapi.yaml
openapi: 3.1.0
info:
  title: kbdb API
  description: >
    API for tracking a personal mechanical keyboard collection: keyboards,
    switches, keycap sets, and the builds that combine them. Routes are
    nested under /v1/users/{userId}/... to support both a caller's own
    collection and items other users have explicitly shared with them —
    see the userId path parameter's description for the exact rules.
  version: "1.0.0"
  contact:
    url: https://mykeebs.info
  license:
    name: MIT
    url: https://github.com/rogueserenity/kbdb/blob/main/LICENSE
servers:
  - url: https://api.mykeebs.info
security:
  - BearerAuth: []
...
paths:
...
  /v1/users/{userId}/switches:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [switches]
      summary: List switches
      operationId: listSwitches
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's switches. Otherwise, lists only that user's switches explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of switches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwitchListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [switches]
      summary: Create a switch
      operationId: createSwitch
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchInput"
      responses:
        "201":
          description: The created switch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Switch"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/switches/{switchId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/SwitchId"
    get:
      tags: [switches]
      summary: Get a switch
      operationId: getSwitch
      security:
        - {}
        - BearerAuth: []
      description: >
        Returns the switch if it belongs to the caller, or if its owner has shared it with the caller. Returns 404 (not 403) if the switch does not exist, is not owned by the caller, and is not shared with the caller — this avoids revealing that an unshared item exists.
      responses:
        "200":
          description: The requested switch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Switch"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    put:
      tags: [switches]
      summary: Update a switch
      operationId: updateSwitch
      description: >
        userId must be the caller's own subject; updating another user's switch always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchInput"
      responses:
        "200":
          description: The updated switch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Switch"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [switches]
      summary: Delete a switch
      operationId: deleteSwitch
      description: >
        userId must be the caller's own subject; deleting another user's switch always returns 404.
        on_delete controls what happens if the switch is still referenced by a build: "block" (the
        default) fails with 409 and lists the blocking build ids; "cascade" deletes the switch and
        every build that references it, returning their ids; "detach" deletes the switch regardless,
        leaving referencing builds with a dangling switches[].switch id.
      parameters:
        - name: on_delete
          in: query
          schema:
            type: string
            enum: [block, cascade, detach]
            default: block
          description: How to handle a switch still referenced by one or more builds.
      responses:
        "200":
          description: Deleted via cascade; lists the builds that were also deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CascadeDeleteResult"
        "204":
          description: Deleted (block with no references, or detach).
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: Blocked by one or more referencing builds (on_delete=block, the default).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/switches/{switchId}/image:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/SwitchId"
    post:
      tags: [switches]
      summary: Set (or replace) a switch's image
      operationId: setSwitchImage
      description: >
        userId must be the caller's own subject; setting a switch's image on
        another user's switch always returns 404. A switch has at most one
        image — calling this again replaces it, no need to delete first.
        Does not upload the image itself: the response is a presigned S3
        PUT URL the client uploads the actual image bytes to directly. The
        presigned URL requires the exact content_type given here on the
        upload — this lets the API reject an unsupported content_type up
        front, before ever issuing a presigned URL, and lets S3 reject an
        upload that doesn't match what was requested.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadRequest"
      responses:
        "201":
          description: >
            An image slot has been allocated; upload the image bytes to
            upload_url (HTTP PUT, with a Content-Type header matching the
            content_type given in the request) to complete it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwitchImageUpload"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [switches]
      summary: Remove a switch's image
      operationId: deleteSwitchImage
      description: >
        userId must be the caller's own subject; removing a switch's image
        on another user's switch always returns 404.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An IdP-issued OIDC JWT (Authorization Bearer <token>).
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: >
        The IdP subject of the item's owner — the `sub` claim of an access
        token, and the `user_id` field of a profile. If this matches the
        authenticated caller's own subject, the request is against their own
        collection. If it names a different user, the request is against
        that user's collection: reads return only items that user has
        explicitly shared with the caller (or, for a specific item, only if
        that item is shared with the caller — otherwise a 404, not a 403,
        to avoid revealing that an item exists but isn't shared); writes
        (create/update/delete) always fail with a 404 in this case, since a
        caller can never modify another user's items. Not a PII/security
        concern to expose in a URL — an IdP subject is an opaque identifier,
        not derived from email or any other identifying attribute.
...
    SwitchId:
      name: switchId
      in: path
      required: true
      schema:
        type: string
      description: Server-generated identifier for the switch.
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response's next_cursor.
  responses:
    ProblemResponse:
      description: An error, formatted per RFC 9457 (Problem Details for HTTP APIs).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundResponse:
      description: The requested resource does not exist, or does not belong to the authenticated user.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ConflictResponse:
      description: >
        The keycap set or build was modified by a concurrent request while
        this one was in flight, and retries were exhausted. Safe to retry
        the request as-is.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details.
      properties:
        type:
          type: string
          format: uri
          example: https://mykeebs.info/errors/validation-error
        title:
          type: string
          example: Validation Error
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        invalid_params:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              reason:
                type: string
              allowed_values:
                type: array
                items:
                  type: string
        blocking_build_ids:
          type: array
          description: >
            Present on a 409 response caused by the item still being
            referenced by one or more builds. Lists the ids of the blocking
            builds.
          items:
            type: string
      required: [type, title, status]
    CascadeDeleteResult:
      type: object
      description: >
        Returned by a delete call made with on_delete=cascade, listing the
        ids of every build that was deleted alongside the requested item.
      properties:
        deleted_build_ids:
          type: array
          items:
            type: string
      required: [deleted_build_ids]
    Money:
      type: number
      format: double
      description: A price in the collection's tracking currency.
      minimum: 0
      maximum: 100000000000000
      example: 123.45
    Currency:
      type: string
      example: USD
      description: >
        An ISO 4217 currency code, from the item owner's
        preferences.currency. Display-only: kbdb stores no currency with a
        price and does no conversion.
    Visibility:
      type: string
      enum: [public, authenticated, private]
      description: >
        Who can read this item via its GET routes (create/update/delete
        always require the caller to be the item's owner, regardless of
        visibility). "public": readable with no authentication at all.
        "authenticated": readable by any signed-in user, but not
        anonymous callers. "private": readable only by the owner. There
        is no per-user allowlist/sharing mechanism beyond these three
        tiers — visibility is not scoped to specific other users.
...
    ImageUploadRequest:
      type: object
      description: >
        Request body for every image-upload endpoint (keyboard, switch,
        profile avatar, keycap kit, build). The upload is two-step: this
        call allocates a slot and returns a presigned S3 PUT URL, then the
        client PUTs the bytes to that URL with a matching Content-Type.
      required: [content_type]
      additionalProperties: false
      properties:
        content_type:
          type: string
          example: image/png
          description: Open vocabulary — validated against the "image_content_type" lookup at request time.
...
    SwitchMaterial:
      type: object
      additionalProperties: false
      properties:
        top_housing:
          type: string
          description: Open vocabulary — validated against the "switch_material" lookup at request time.
        bottom_housing:
          type: string
          description: Open vocabulary — validated against the "switch_material" lookup at request time.
        stem:
          type: string
          description: Open vocabulary — validated against the "switch_material" lookup at request time.
    SwitchForce:
      type: object
      additionalProperties: false
      properties:
        actuation:
          type: number
          format: double
          description: Actuation force in grams (nominal value).
        bottom_out:
          type: number
          format: double
          description: Bottom-out force in grams (nominal value).
    SwitchSpring:
      type: object
      additionalProperties: false
      properties:
        material:
          type: string
          description: Open vocabulary — validated against the "switch_spring_material" lookup at request time.
        pre_travel:
          type: number
          format: double
          description: Pre-travel distance in mm.
        total_travel:
          type: number
          format: double
          description: Total travel distance in mm.
    SwitchPurchaseInput:
      type: object
      description: >
        Same purchase-lifecycle tracking as PurchaseInput, plus a quantity
        field since switches are bought in bulk.
      additionalProperties: false
      properties:
        vendor:
          type: string
          description: Open vocabulary — validated against the "vendor" lookup at request time.
        price:
          $ref: "#/components/schemas/Money"
        order_date:
          type: string
          format: date
        delivery_date:
          type: string
          format: date
        order_status:
          type: string
          description: Open vocabulary — validated against the "order_status" lookup at request time.
        quantity:
          type: integer
    SwitchPurchase:
      type: object
      description: >
        The read form of SwitchPurchaseInput.
      properties:
        vendor:
          type: string
        price:
          $ref: "#/components/schemas/Money"
        currency:
          $ref: "#/components/schemas/Currency"
          description: >
            The owner's display currency, for labelling price.
            Present exactly when price is.
        order_date:
          type: string
          format: date
        delivery_date:
          type: string
          format: date
        order_status:
          type: string
        quantity:
          type: integer
    SwitchInput:
      type: object
      required: [brand, name, type, visibility]
      additionalProperties: false
      properties:
        brand:
          type: string
          minLength: 1
          pattern: "\\S"
        manufacturer:
          type: string
        name:
          type: string
          minLength: 1
          pattern: "\\S"
        type:
          type: string
          description: Open vocabulary — validated against the "switch_type" lookup at request time.
        pins:
          type: integer
          example: 5
          description: PCB mount pin count.
        factory_lubed:
          type: boolean
        material:
          $ref: "#/components/schemas/SwitchMaterial"
        force:
          $ref: "#/components/schemas/SwitchForce"
        spring:
          $ref: "#/components/schemas/SwitchSpring"
        purchase:
          $ref: "#/components/schemas/SwitchPurchaseInput"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
    SwitchImage:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: >
            A freshly-minted, short-lived presigned S3 GET URL. Generated
            new on every fetch — do not cache or persist this value, it
            expires. A switch has at most one image (its vendor render/
            product shot) — see POST/DELETE .../switches/{switchId}/image.
    SwitchImageUpload:
      type: object
      required: [upload_url]
      properties:
        upload_url:
          type: string
          format: uri
          description: >
            Presigned S3 PUT URL. Short-lived — upload promptly.
    Switch:
      type: object
      required: [id, brand, name, type]
      properties:
        id:
          type: string
        brand:
          type: string
        manufacturer:
          type: string
        name:
          type: string
        type:
          type: string
        pins:
          type: integer
          example: 5
          description: PCB mount pin count.
        factory_lubed:
          type: boolean
        material:
          $ref: "#/components/schemas/SwitchMaterial"
        force:
          $ref: "#/components/schemas/SwitchForce"
        spring:
          $ref: "#/components/schemas/SwitchSpring"
        purchase:
          $ref: "#/components/schemas/SwitchPurchase"
        notes:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this switch. Only ever present for the
            switch's owner - no preference can expose it to another
            caller.
        image:
          $ref: "#/components/schemas/SwitchImage"
          description: >
            Managed via the dedicated image endpoint, not via PUT on
            the switch itself — see POST/DELETE
            .../switches/{switchId}/image.
    SwitchSummary:
      type: object
      properties:
        id:
          type: string
        brand:
          type: string
        name:
          type: string
        type:
          type: string
          description: Open vocabulary — validated against the "switch_type" lookup at request time.
        order_status:
          type: string
        image:
          $ref: "#/components/schemas/SwitchImage"
          description: A freshly-minted presigned URL to the switch's image, if it has one set.
        price:
          $ref: "#/components/schemas/Money"
          description: >
            What the switch cost. Present for the owner only if their
            show_price_to_me preference is set, and for any other caller
            only if the owner's show_price_to_others is.
        currency:
          $ref: "#/components/schemas/Currency"
          description: >
            The owner's display currency, for labelling price.
            Present exactly when price is.
        visibility:
          $ref: "#/components/schemas/Visibility"
          description: >
            Who can read this switch. Only ever present for the
            switch's owner - no preference can expose it to another
            caller.
    SwitchListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SwitchSummary"
        next_cursor:
          type: [string, "null"]
...